Add pointer initialization in MPU6050.cpp
The uint8_t* verifyBuffer was uninitialized. I don't think this causes any problems, since 
it is only checked if verify is true, and that condition does initialize verifyBuffer elsewhere
in the code. However, it's probably safer to initialize it to 0 or nullptr in case some future
edit does something different.